requirements.txt
netmiko==3.3.2
PyYAML==5.3.1